refactor: extract client as single library
