+++ .env
@@ -1,0 +1,4 @@
+YASHANDB_HOST=127.0.0.1
+YASHANDB_PORT=1688
+YASHANDB_USERNAME=sys
+YASHANDB_PASSWORD=yasdb_123
