+++ yasqlplus-client/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "yasqlplus-client"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "1.0.50"
+
+[build-dependencies]
+bindgen = "0.69.1"
