+++ yasqlplus-client/src/native.rs
