+++ yasqlplus-client/src/wrapper/error.rs
@@ -2,10 +2,13 @@ use super::{get_error, DiagInfo};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
-    #[error("{0}")]
+    #[error("{0:?}")]
     YasClient(DiagInfo),
-    #[error(transparent)]
-    Other(#[from] anyhow::Error),
+
+    // #[error(transparent)]
+    // Other(#[from] Box<dyn std::error::Error>),
+    #[error("Unknown error")]
+    Other,
 }
 
 impl Error {
