+++ yasqlplus-client/src/wrapper/result.rs
