+++ yasqlplus-client/src/wrapper/row.rs
